+++ src/GRA.Data/Repository/RoleRepository.cs
@@ -1,10 +1,13 @@
-﻿using GRA.Domain.Repository;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using Microsoft.Extensions.Logging;
-using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
+using AutoMapper.QueryableExtensions;
+using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
+using GRA.Domain.Repository;
+using GRA.Domain.Repository.Extensions;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace GRA.Data.Repository
 {
@@ -15,33 +18,132 @@ namespace GRA.Data.Repository
             ILogger<RoleRepository> logger) : base(repositoryFacade, logger)
         { }
 
-        public async Task AddPermissionAsync(int userId, string name)
+        public async Task<IEnumerable<Role>> GetAllAsync()
         {
-            await _context.Permissions.AddAsync(new Model.Permission
-            {
-                Name = name,
-                CreatedBy = userId,
-                CreatedAt = _dateTimeProvider.Now
-            });
+            return await DbSet
+                .AsNoTracking()
+                .ProjectTo<Role>()
+                .ToListAsync();
         }
 
-        public async Task AddPermissionToRoleAsync(int userId, int roleId, string permissionName)
+        public async Task<DataWithCount<IEnumerable<Role>>> PageAsync(BaseFilter filter)
         {
-            var permission = await _context.Permissions
-                .Where(_ => _.Name == permissionName)
-                .SingleOrDefaultAsync();
-            if (permission == null)
-            {
-                throw new Exception($"Permission '{permissionName}' not found.");
-            }
+            var roles = DbSet.AsNoTracking();
+            var count = await roles.CountAsync();
+            var data = await roles
+                .OrderBy(_ => _.Name)
+                .ApplyPagination(filter)
+                .ProjectTo<Role>()
+                .ToListAsync();
 
-            await _context.RolePermissions.AddAsync(new Model.RolePermission
+            return new DataWithCount<IEnumerable<Role>>
             {
-                RoleId = roleId,
-                PermissionId = permission.Id,
-                CreatedBy = userId,
-                CreatedAt = _dateTimeProvider.Now
+                Data = data,
+                Count = count
+            };
+        }
+
+        public async Task<Role> AddSaveAsync(int userId, Role role, IEnumerable<string> permissions)
+        {
+            var now = _dateTimeProvider.Now;
+            var roleEntity = _mapper.Map<Role, Model.Role>(role);
+
+            var rolePermissions = _context.Permissions
+                .AsNoTracking()
+                .Where(_ => permissions.Contains(_.Name))
+                .Select(_ => new Model.RolePermission
+                {
+                    Role = roleEntity,
+                    PermissionId = _.Id,
+                    CreatedAt = now,
+                    CreatedBy = userId
+                });
+            await _context.RolePermissions.AddRangeAsync(rolePermissions);
+            await _context.SaveChangesAsync();
+            return _mapper.Map<Model.Role, Role>(roleEntity);
+        }
+
+        public async Task UpdateSaveAsync(int userId, Role role, List<string> permissionsToAdd,
+            List<string> permissionsToRemove)
+        {
+            await UpdateAsync(userId, role);
+
+            var now = _dateTimeProvider.Now;
+
+            var addPermissions = _context.Permissions
+                .AsNoTracking()
+                .Where(_ => permissionsToAdd.Contains(_.Name))
+                .Select(_ => new Model.RolePermission
+                {
+                    RoleId = role.Id,
+                    PermissionId = _.Id,
+                    CreatedAt = now,
+                    CreatedBy = userId
+                });
+            var removePermissions = _context.RolePermissions
+                .Where(_ => _.RoleId == role.Id && role.IsAdmin == false
+                    && permissionsToRemove.Contains(_.Permission.Name));
+
+            await _context.RolePermissions.AddRangeAsync(addPermissions);
+            _context.RolePermissions.RemoveRange(removePermissions);
+            await _context.SaveChangesAsync();
+        }
+
+        public override async Task RemoveSaveAsync(int userId, int roleId)
+        {
+            var rolePermissions = _context.RolePermissions.Where(_ => _.RoleId == roleId);
+            _context.RolePermissions.RemoveRange(rolePermissions);
+
+            var userRoles = _context.UserRoles.Where(_ => _.RoleId == roleId);
+            _context.UserRoles.RemoveRange(userRoles);
+
+            await RemoveAsync(userId, roleId);
+
+            await _context.SaveChangesAsync();
+        }
+
+        public async Task<IEnumerable<string>> GetAllPermissionsAsync()
+        {
+            return await _context.Permissions
+                .AsNoTracking()
+                .Select(_ => _.Name)
+                .OrderBy(_ => _)
+                .ToListAsync();
+        }
+
+        public async Task AddPermissionListAsync(IEnumerable<string> names)
+        {
+            var now = _dateTimeProvider.Now;
+
+            var permissions = names.Select(_ => new Model.Permission
+            {
+                Name = _,
+                CreatedAt = now,
+                CreatedBy = -1
             });
+
+            var rolePermissions = DbSet.AsNoTracking()
+                .Where(_ => _.IsAdmin)
+                .SelectMany(_ => permissions, (r, p) => new Model.RolePermission
+                {
+                    RoleId = r.Id,
+                    Permission = p,
+                    CreatedAt = now,
+                    CreatedBy = -1
+                });
+
+            await _context.RolePermissions.AddRangeAsync(rolePermissions);
+        }
+
+        public void RemovePermissionList(IEnumerable<string> names)
+        {
+            var permissions = _context.Permissions.Where(_ => names.Contains(_.Name));
+
+            var rolePermissions = _context.RolePermissions
+                .Where(_ => permissions.Select(p => p.Id).Contains(_.PermissionId));
+
+            _context.RolePermissions.RemoveRange(rolePermissions);
+            _context.Permissions.RemoveRange(permissions);
         }
 
         public async Task<IEnumerable<string>> GetPermisisonNamesForUserAsync(int userId)
@@ -61,6 +163,34 @@ namespace GRA.Data.Repository
                 .Distinct()
                 .OrderBy(_ => _)
                 .ToListAsync();
+        }
+
+        public async Task<IEnumerable<string>> GetPermissionNamesForRoleAsync(int roleId)
+        {
+            return await _context.RolePermissions
+                .AsNoTracking()
+                .Where(_ => _.RoleId == roleId)
+                .Select(_ => _.Permission.Name)
+                .OrderBy(_ => _)
+                .ToListAsync();
+        }
+
+        public async Task<bool> ListContainsAdminRoleAsync(IEnumerable<int> roleIds)
+        {
+            return await DbSet
+                .AsNoTracking()
+                .Where(_ => _.IsAdmin && roleIds.Contains(_.Id))
+                .AnyAsync();
+        }
+
+        public async Task<int> GetUsersWithAdminRoleCountAsync()
+        {
+            return await _context.UserRoles
+                .AsNoTracking()
+                .Where(_ => _.Role.IsAdmin)
+                .Select(_ => _.UserId)
+                .Distinct()
+                .CountAsync();
         }
     }
 }