+++ src/GRA.Domain.Model/Role.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace GRA.Domain.Model
 {
@@ -7,5 +8,7 @@ namespace GRA.Domain.Model
         [Required]
         [MaxLength(255)]
         public string Name { get; set; }
+        public bool IsAdmin { get; set; }
+        public int PermissionCount { get; set; }
     }
 }