+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Documentation in docs/ folder to update the manual for version 4
 - Metadata description and structured data for Events
 - Ability to restrict triggers from activating until a specified date
+- Role management
+- Authorization code management
+- Participant/role assignment
+- Automatically assign any new permissions to the System Administrator role
 
 ### Fixed
 - Broken household URLs in Mission Control (leading spaces)